[fix] 타입 오류 수정
bucketId 가 Integer인데 long이라 오류 발생하여 재수정

+++ ToBeBucketBackend/src/main/java/com/example/ToBeBucket/Entity/Bucket.java
@@ -16,7 +16,7 @@ public class Bucket {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "bucketId", nullable = false)
-    private Long bucketId;
+    private Integer bucketId;
 
     @Column(name = "userId", length = 255, nullable = false)
     private String userId;
